+++ Chrominsky.Utils/Models/Base/BaseEntity.cs
@@ -51,6 +51,6 @@ public abstract class BaseEntity
     /// </summary>
     [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
     [Required]
-    [Newtonsoft.Json.JsonConverter(typeof(JsonStringEnumConverter))]
+    [System.Text.Json.Serialization.JsonConverter(typeof(JsonStringEnumConverter))]
     public DatabaseEntityStatus? Status { get; set; }
 }